### 2022-09-02 #56 @simonvh
Presubmission enquiry for genomepy
Submitting Author: Simon van Heeringen (@simonvh)
Package Name: genomepy
One-Line Description of Package: simple and straightforward downloading and management of genomic data.
Repository Link (if existing):   https://github.com/vanheeringen-lab/genomepy

---

Genomepy is designed to provide a simple and straightforward way to download and use genomic data (genome sequences and gene annotation). This includes (1) searching available data, (2) showing the available metadata, (3) automatically downloading, preprocessing and matching data and (4) generating optional aligner indexes. All with sensible, yet controllable defaults. Currently, genomepy supports Ensembl, UCSC, NCBI and GENCODE.

Genomepy streamlines **data retrieval** from genome repositories (for example `genomepy install hg38 --annotation` instead of searching the genome download link by hand, and downloading gene annotation from a separate source. It can read various formats and converts gene annotation to common and consistent format (data **extraction** and **munging**). All steps, download sources, identifiers are logged for **reproducibility**.

Any computational biologist or bioinfomatician in genomics. Genomepy makes it very easy to work with different genomes and to incorporate genome download in pipelines. Even when only working with one species (ie human or mouse), it streamlines genome download and use.

There are three that we are aware of: [GoGetData](https://www.nature.com/articles/s41467-021-22381-z) and [RefGenie](http://refgenie.databio.org/en/latest/) can download genomes and genomic assets. However, these packages depend on predefined recipes, while genomepy can download any genome that is hosted on one of the major providers. In addition, genomepy provides a Python API to use the genomic data. In addition [ncbi-genome-download](https://github.com/kblin/ncbi-genome-download) focusses on bacterial and fungal genomes. Genomepy is agnostic to species.

An older version of Genomepy was already [published in JOSS](https://joss.theoj.org/papers/10.21105/joss.00320).
