### 2022-09-09 #56 @NickleDave
Hi @simonvh, welcome to pyOpenSci and thank you so much for your detailed presubmission inquiry!

Genomepy definitely looks in scope to me.

There is a bit of a question about how to handle the fact that it's already been reviewed by JOSS.
In this case I see that the JOSS review was in 2017, and that you all have been steadily developing the package since that time.

Can you tell us a little bit more about what you are hoping to get out of a pyOpenSci review?

If it is something like "I feel like we would benefit from a new review after 5 years of development" then I think it could definitely be appropriate.

For sure I do not mean to discourage you.
It does seem like Genomepy would be a good fit and I am fairly confident we can find editors and reviewers from previous reviews of bioinformatics packages.

We are in a better position to put more process around this situtation now that our executive director @lwasser has been able to move pyOpenSci to a new fiscal host (we're just starting up reviews again after a pause). This is a good time for us to figure out how to handle it, if you can bear with us as we work it out on the fly 😄.

Genomepy looks like a great library and we'd like to help you if we can.
