### 2022-09-09 #56 @lwasser
Thank you @NickleDave !! @simonvh yup, we just want to understand your goals in submitting here. I think it would be great to support packages that have been reviewed by JOSS given we offer more python specific feedback in our reviews in addition to providing visibility of being a pyOS vetted package! So there are many wins here. But any insight into your goals for this review would be wonderful as we work through how we handle this type of submission! We look forward to hearing from you!
