### 2022-09-14 0626a87
Tests (#219)
* Update __init__.py

* Update __init__.py

* fix tests

* Update CHANGELOG.md

Co-authored-by: Maarten-vd-Sande <Maarten-vd-Sande@users.noreply.github.com>
