### 2022-09-22 #56 @NickleDave
Hi again @simonvh, just following up here.

@lwasser and I had a chance to talk a little bit more about this presubmission today.

We do think in this case there's no problem with the previous review from JOSS, given that it was 2017 and it does look like you all have done significant development of the package since that time.

I am confident we can find appropriate editors and reviewers by reaching out contributors involved with previous bioinformatics and -omics packages we have reviewed.

Please feel free to start a new issue for a full submission--once you do we can close this one--and just let us know if you have any questions.
