### 2022-09-23 #56 @simonvh
Hi @NickleDave and @lwasser, sorry for the delay in answering! Thanks for the input here. Just to add to the discussion: my aim was primarily to 1) gain more visibility for genomepy and 2) to support pyOpenSci, which I didn't know about and heard about through Twitter. In addition to considering pyOpenSci, we have also recently submitted genomepy as a journal paper (preprint here: https://arxiv.org/abs/2209.00842).

From your reactions, it seems that the review process is more extensive than I had thought, which is good news as it would focus more on the technical side and we're always eager to learn there. We have indeed significantly expanded on genomepy as originally published in JOSS. If you indeed think it would be OK from your side, then I would be happy to start a full submission.
