### 2022-09-23 #56 @NickleDave
Thank you @simonvh.

We're glad to help gain visibility for `genomepy` and appreciate your support.

>  [I]t seems that the review process is more extensive than I had thought. We have indeed significantly expanded on genomepy as originally published in JOSS.

Great, It sounds like we're on the same page that a review would be appropriate given the amount of time that's passed and the development that's been done :+1:

And just to clarify what pyOpenSci does that is different:
* We drive adoption of community standards and emphasize robust Python-specific review
* We curate a list of vetted tools - that people can trust will have some level of maintenance
* We will check in on maintenance and sunset / communicate as we need to or find a new maintainer
