### 2022-09-23 #56 @NickleDave
> In addition to considering pyOpenSci, we have also recently submitted genomepy as a journal paper (preprint here: https://arxiv.org/abs/2209.00842).

Ok understood.
We have had pre-prints submitted before (see e.g. #26 ) but I think this is the first time someone has simultaneously submitted to pyOpenSci and a journal.

We just want to be sure we're handling this the right way.
@lwasser is checking in with contacts at JOSS and rOpenSci.

There's also a question of whether people would cite the journal article, the version of the package that pyOpenSci reviewed, or both.
(Both can work! If you use a [CITATION.cff](https://citation-file-format.github.io/) :slightly_smiling_face:)
