### 2022-09-28 #56 @NickleDave
Hi @simonvh, one more update on this.
In regards to publishing in other venues, we will be adopting the approach that rOpenSci takes, to avoid potential conflicts. From their guide:

> .1.1 Publishing in other Venues
>
>     We strongly suggest submitting your package for review before publishing on CRAN or submitting a software paper describing the package to a journal. Review feedback may result in major improvements and updates to your package, including renaming and breaking changes to functions. We do not consider previous publication on CRAN or in other venues sufficient reason to not adopt reviewer or editor recommendations.
>     Do not submit your package for review while it or an associated manuscript is also under review at another venue, as this may result on conflicting requests for changes.

I apologize that we did not previously have clear guidance in place. @lwasser is in the process of adding this language to our guidebook now.

Long story short:
We really appreciate that you want to support pyOpenSci. We would definitely be more than happy to provide you with a review **after** the review process ends at the journal where you have currently submitted, especially if we can help you add a shiny badge to your repo and get more visibility for `genomepy`.

Please let me know if that's clear. I appreciate your understanding and patience as we figure out this process on the fly.
