### 2022-09-29 #56 @simonvh
HI @NickleDave, thanks and no worries! I understand that this all needs to be figured out. We'll wait, and I'll check back in at a later stage.

Just as an aside, I would personally love for venues like JOSS and pyOpenSci to also be sufficient as publication. At least going by the JOSS review, the review process is much more useful, technically informative and geared towards improving a software package. Less adversarial than the traditional review process as well. However, as far as visibility goes, it is (sadly) in the biosciences still no match for a Pubmed-indexed journal. So in that sense, double citations would I guess still be the way to go.

I'll close the thread, we can always re-open in the future. Thanks for thinking along!
