### 2022-09-29 #56 @lwasser
@simonvh thank you so much for this feedback. I agree. RopenSci actually has agreements with several publications to allow their review to suffice for that publication's review. Then the publication can focus on the paper part and we focus on the technical part. Please feel free to mention this to the publication that you are working with as we are open to it. We have that type of agreement with JOSS - our review is much more python specific and we offer more things to maintainers such as supporting longer term maintenance and visibility in the scientific python community. So JOSS accepts our review as a part of their process if the package is in scope for them.

SO in short - bare with us as we figure things out. we've been running peer review for a few years but are just getting organized as a new independent organization. As such we have many partnerships to make and much to do!

You are WELCOME to submit to us after your review is done if you want ANOTHER review and the visibility of our platform behind you. And in the meantime send whomever our way to start discussions around the double review vs just accepting our reviews!  leah at pyopensci.org is a good way to reach me if you have any other questions or thoughts as well!
